Fixed tree->list conversion which were wrong
They returned '[] for an empty list, not #nil. This
is a problem because '[] is a list with a single item.
nil. Which is NOT the same as an empty list. This led to
crashes in WolkenWelten when trying to iterate over an empty
event handler, calling all the items in that tree, and due to
this error, it tried to eval nil, which led to a crash.