add hash table over symbol table array
This adds a hash table based lookup for symbols.  The existing structure
for symbols remains unchanged, and the hash table is merely added on
top.  That is, lSymbolList stores the actual symbols, lSymbolIndex
stores positions in lSymbolList.  lSymbolBackIndex goes in the opposite
direction, mapping positions in lSymbolList to positions in lSymbolIndex
which is needed to properly clean up the index in lSymbolFree.
lSymbolIndex can get away with a single int per slot in the hash table.
If the value is == 0, the slot is empty, if it's < 0 the symbol in the
slot was freed (and the slot can be reused), if it's > 0 the symbol in
the slot is in use.
In total lSymbolIndex and lSymbolBackIndex store 2 * SYM_MAX ints,
meaning they need 131kb of memory.