Added support for late binding in Bytecode λs
This is necessary since serialized bytecode λs contain
references to other functions, either written in Nujel or
Native ones, since we can't include their entire definition in that
serialized string we instead just store the symbol, then when we eval
the bytecoded λ for the first time we iterate over all instructions and
exchange resolve all symbols. We could have also immediatly resolved
the symbol in the reader but this would inhibit mutual recursion,
whereas this scheme allows for it since the Pointers don't change.